+++ src/pages/coimbatore/partner-with-us.tsx
@@ -1,0 +1,6 @@
+import Partner from "@/components/Partner";
+import React, { FC } from "react";
+
+const PartnerWithUs: FC = () => <Partner />;
+
+export default PartnerWithUs;